requirements.txt
blinker==1.9.0
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-WTF==1.2.2
fonttools==4.58.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
narwhals==1.39.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
plotly==6.0.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
WTForms==3.2.1
